restricted clear to mods and admins
%help%cmd still not working past %help%help

+++ index.js
@@ -224,6 +224,9 @@ bot.on('message', message => {
     }
 
     if (command === commands[6]) {
+      let modRole === message.guild.roles.find("name", "Moderator");
+      let adminRole === messsage.guild.roles.find("name", "Admin");
+      if(!message.member.roles.has(modRole.id || adminRole.id) || message.member.id === message.guild.ownerID) {
         if (isNaN(args.join(" "))) {
             deleteBotMsg("That's not a valid number.")
 
@@ -253,6 +256,10 @@ bot.on('message', message => {
             message.reply(":8ball: | I can't read that, try something else.")
             return
         }
+      } else {
+        deleteBotMsg("You don't have the sufficient permissions.")
+        return;
+      }
     }
 
     if (command === commands[8]) {
